echo done, dollar_sign_handler done, finishing wildcards
