+++ .vscode/settings.json
@@ -55,11 +55,13 @@
   "C_Cpp_Runner.useLeakSanitizer": false,
   "C_Cpp_Runner.showCompilationTime": false,
   "C_Cpp_Runner.useLinkTimeOptimization": false,
-"C_Cpp_Runner.msvcSecureNoWarnings": false,
-"files.associations": {
+  "C_Cpp_Runner.msvcSecureNoWarnings": false,
+  "files.associations": {
 	"cstring": "c",
 	"libft.h": "c",
 	"ft_builtins.h": "c",
-	"minishell.h": "c"
+	"minishell.h": "c",
+	"wildcard.h": "c",
+	"format": "c"
 }
 }
