+++ src/wildcards/ft_wildcard.h
@@ -1,0 +1,6 @@
+#ifndef FT_WILDCARD_H
+# define FT_WILDCARD_H
+
+# include "../minishell.h"
+
+#endif
